Move SlackMessage to own module
Also simplify joins/parts a bit
